requirements.txt
Chameleon==2.14
FormEncode==1.2.6
Mako==0.9.1
MarkupSafe==0.18
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.1
SimpleCRUD==0.0
WTForms==1.0.5
WebHelpers==1.3
WebOb==1.3.1
anykeystore==0.2
apex==0.9.10dev
argparse==1.2.1
cryptacular==1.4.1
oauthlib==0.6.1
pbkdf2==1.3
psycopg2==2.5.2
pyramid==1.4.5
pyramid-debugtoolbar==1.0.9
pyramid-mailer==0.13
pyramid-mako==0.3.1
pyramid-simpleform==0.6.1
pyramid-tm==0.7
python-openid==2.2.5
repoze.lru==0.6
repoze.sendmail==4.1
requests==2.2.1
requests-oauthlib==0.4.0
transaction==1.4.1
translationstring==1.1
velruse==1.1.1
venusian==1.0a8
waitress==0.8.8
wsgiref==0.1.2
wtforms-recaptcha==0.3.0
zope.deprecation==4.1.0
zope.interface==4.0.5
zope.sqlalchemy==0.7.4
